5.3 Upgrade, Multiplayer InstancedStruct
- Upgraded to 5.3, which adds support for replicating and RPCing FInstancedStructs!

+++ Source/SaiyoraV4/Public/CoreClasses/SaiyoraGameState.h
@@ -26,7 +26,7 @@ public:
 
 	ASaiyoraGameState();
 	virtual void Tick(float DeltaSeconds) override;
-	virtual float GetServerWorldTimeSeconds() const override { return WorldTime; }
+	virtual double GetServerWorldTimeSeconds() const override { return WorldTime; }
 	void UpdateClientWorldTime(const float ServerTime) { WorldTime = ServerTime; }
 
 private:
